Refactor wolfssl integration
