+++ tlspuffin/Cargo.toml
@@ -47,7 +47,7 @@ sancov_libafl = ["libafl_targets/sancov_pcguard_hitcounts"]
 asan = ["openssl/asan"]
 
 # WolfSSL bindings are used
-wolfssl-binding = ["wolfssl-sys", "libc"]
+wolfssl-binding = ["wolfssl-sys", "libc", "foreign-types-shared", "foreign-types", "bitflags"]
 # OpenSSL bindings are used
 openssl-binding = ["openssl", "foreign-types-shared"]
 
@@ -80,6 +80,8 @@ security-claims = { path = "../tlspuffin-claims", optional = false }
 
 # WolfSSL
 wolfssl-sys = { path = "../deps/wolfssl-sys", features = ["sancov"], optional = true }
+foreign-types = { version = "0.5.0", optional = true }
+bitflags = { version = "1.3.2", optional = true }
 libc = { version = "0.2.126", optional = true }
 
 # LibAFL
