+++ tlspuffin/src/wolfssl/util.rs
@@ -1,6 +1,8 @@
+use crate::error::Error;
 use libc::c_int;
+use std::io::ErrorKind;
 
-use crate::wolfssl::error::ErrorStack;
+use crate::wolfssl::error::{Error as SSLError, ErrorStack};
 
 pub fn cvt_p<T>(r: *mut T) -> Result<*mut T, ErrorStack> {
     if r.is_null() {
